Minor fixes and start adding tests
The major change is the start of the PDF tests.

New round-trip function requires separate
github static_pdfs repository.  This repo has
a lot of challenging PDFs in it that are not yet
done right.

Current round-trip tests do not yet have correct
automated pass/fail flags, but they do output PDFs
that can be examined.

Minor fixes include the fix for the watermarking
program (issue #12) and the ability to insure
that floats are never represented in exponent
notation.