Merge in changes from master
Also refactor, optimize a few things, and make it clear what the
algorithm differences are between this branch and the master.

It currently works like the prior nerijus branch (on the testcases)
other than a couple of fixes brought over from the master branch.

There is one failing testcase, due to the rewrite of the trailer
dict.

There is also a feature_stream_objects switch that makes it easy
to change the behavior over to the current master branch behavior
(and fix the failing testcase).

The primary optimization was using hexlify instead of struct, which
should speed up the compressed reference stream reading.

So, thanks to Nerijus, we have a very good starting point going
forward for PDFs with compressed object streams, and it is now
easy to see what the differences are.