rotations don't work in Py3
In Python 3, `float / int == float`, so the return value from `get_rotation` will usually be a float.  In `rotate_point`, however, `float & int` is a `TypeError`, so `rotation & 1` fails when `rotation` is a float.  Therefore, for any rotation that is valid in `get_rotation`, `rotate_rect` will raise a `TypeError`.