Refactor stream object table reading
I'm not sure why it was checking whether the table entries were numeric,
but it may have to do with the fact that the stream floc member would not
be properly set up until the first call to next().  In any case, modify
the code to properly set up the floc member, and then simplify the loop.