When replacing pages in pdfwriter, update objects that refer to the old page.
This accomplishes the same thing as abrasive's PR #60.  That PR worked for all the
test_roundtrip.py testcases, but failed on the reportlab test cases (made PDFs that
were not right).  I have not examined the failure too deeply, but believe it may
have been because the underlying page objects were freed and their object IDs were
reused.

In any case, there was already code in pdfrw to accomplish something very similar for
the parents of the pages, so I modified that slightly to support changing the pages,
as well.