Fix PdfWriter dangling object references.
This attacks the same problem as some crufty old code in pdfrw, plus
the attempted fixes in PR #60 and PR #82.  Those fixes worked on
many PDFs, but attempted to do the impossible.  If we're building a
new page hierarchy, then we really do need a new hierarchy.

This fix removes any annotations and beads (which could wend deep
into an old page hierarchy) from any added page which is part of
an old hierarchy.  We detect the old page hierarchy by the presence
of the /Parent entry in the page dictionary.

If someone wants to preserve annotations, then we'll have to write
code to do that properly.

I am not 100% happy with making PdfWriter assume that /B and /Annots
are the only dict entries that could cause cycles, but so far, the only
PDFs I have seen where this is not true were completely broken to start
with.